refactor: button (#40077)
* fix: type updated from any, and comments removed #15930

* refactor: removed unnecessary comments, created helper for button

* fix: cleanupTimer

* fix: import fixed

* Update components/button/button.tsx


* Update components/dropdown/dropdown-button.tsx
