+++ components/button/button-group.tsx
@@ -21,12 +21,10 @@ const ButtonGroup: React.FC<ButtonGroupProps> = (props) => {
   const { prefixCls: customizePrefixCls, size, className, ...others } = props;
   const prefixCls = getPrefixCls('btn-group', customizePrefixCls);
 
-  // Here we only need hashId
   const [, , hashId] = useToken();
 
-  // large => lg
-  // small => sm
   let sizeCls = '';
+
   switch (size) {
     case 'large':
       sizeCls = 'lg';
