+++ components/button/index.ts
@@ -1,7 +1,9 @@
 import Button from './button';
 
 export type { SizeType as ButtonSize } from '../config-provider/SizeContext';
-export type { ButtonProps, ButtonShape, ButtonType } from './button';
+export type { ButtonProps } from './button';
 export type { ButtonGroupProps } from './button-group';
+
+export * from './buttonHelpers';
 
 export default Button;
