+++ components/dropdown/dropdown-button.tsx
@@ -1,16 +1,16 @@
-import EllipsisOutlined from '@ant-design/icons/EllipsisOutlined';
 import classNames from 'classnames';
 import * as React from 'react';
-import type { ButtonProps } from '../button';
+import EllipsisOutlined from '@ant-design/icons/EllipsisOutlined';
 import Button from '../button';
-import type { ButtonHTMLType } from '../button/button';
-import type { ButtonGroupProps } from '../button/button-group';
 import { ConfigContext } from '../config-provider';
-import { useCompactItemContext } from '../space/Compact';
-import type { DropdownProps } from './dropdown';
-import Dropdown from './dropdown';
 import Space from '../space';
+import { useCompactItemContext } from '../space/Compact';
+import Dropdown from './dropdown';
 import useStyle from './style';
+
+import type { ButtonProps, ButtonHTMLType } from '../button';
+import type { ButtonGroupProps } from '../button/button-group';
+import type { DropdownProps } from './dropdown';
 
 export type DropdownButtonType = 'default' | 'primary' | 'ghost' | 'dashed' | 'link' | 'text';
